Fix dependency options to specify `master` git branch
A newer version of Mix has begun to fail when running `deps.get`
while a suitable version of Slime is not already in the `/deps`
directory. This patch presumes that the head of the default
branch (in this case, presumed to be `master`) of the current
fork contains the required target version of Slime for the project.